+++ pallets/governance/src/benchmarking.rs
@@ -27,29 +27,28 @@ fn submit_dao_application<T: Config>() -> Result<(), &'static str> {
 fn register_mock<T: Config>(
     key: T::AccountId,
     module_key: T::AccountId,
-    stake: u64,
     name: Vec<u8>,
 ) -> Result<(), &'static str> {
     let address = "test".as_bytes().to_vec();
     let network = "testnet".as_bytes().to_vec();
-    BurnConfig::<T>::mutate(|cfg| cfg.min_burn = 0);
+
+    let enough_stake = 10000000000000u64;
     SubspaceMod::<T>::add_balance_to_account(
         &key,
-        SubspaceMod::<T>::u64_to_balance(stake + SubnetBurn::<T>::get() + 2000).unwrap(),
+        SubspaceMod::<T>::u64_to_balance(SubnetBurn::<T>::get() + enough_stake).unwrap(),
     );
     let metadata = Some("metadata".as_bytes().to_vec());
     SubspaceMod::<T>::register(
-        RawOrigin::Signed(key).into(),
+        RawOrigin::Signed(key.clone()).into(),
         network,
         name,
         address,
-        stake.into(),
-        module_key,
+        module_key.clone(),
         metadata,
     )?;
+    SubspaceMod::<T>::increase_stake(&key, &module_key, enough_stake);
     Ok(())
 }
-
 benchmarks! {
     //---------------------------------
     //Adding proposals
@@ -88,7 +87,8 @@ benchmarks! {
             params.governance_config.proposal_expiration,                // proposal_expiration: the block number, proposal expires at
             params.general_subnet_application_cost,     // general_subnet_application_cost
             params.kappa,
-            params.rho
+            params.rho,
+            params.subnet_immunity_period
         )
 
 
@@ -97,7 +97,7 @@ benchmarks! {
         let caller: T::AccountId = account("Alice", 0, 1);
 
         // register the subnet
-        register_mock::<T>(caller.clone(), caller.clone(), 100000000000000u64 + SubnetBurn::<T>::get(),
+        register_mock::<T>(caller.clone(), caller.clone(),
     "test".as_bytes().to_vec())?;
         let netuid = SubspaceMod::<T>::get_netuid_for_name("testnet".as_bytes()).unwrap();
 
@@ -173,7 +173,7 @@ benchmarks! {
     add_subnet_custom_proposal {
         let caller: T::AccountId = account("Alice", 0, 1);
         // The subnet has to exist
-        register_mock::<T>(caller.clone(), caller.clone(), 100000000000000u64,
+        register_mock::<T>(caller.clone(), caller.clone(),
     "test".as_bytes().to_vec())?;     // Add alice fund to submit the proposal
         SubspaceMod::<T>::add_balance_to_account(&caller,
     SubspaceMod::<T>::u64_to_balance(1_000_000_000_000_000).unwrap());
@@ -204,7 +204,7 @@ benchmarks! {
     vote_proposal {
         let caller: T::AccountId = account("Alice", 0, 1);
         // Register Alice such that she has funds to vote
-        register_mock::<T>(caller.clone(), caller.clone(), 100000000000000u64, "test".as_bytes().to_vec())?;
+        register_mock::<T>(caller.clone(), caller.clone(), "test".as_bytes().to_vec())?;
 
         // Add Alice's funds to submit the proposal
         SubspaceMod::<T>::add_balance_to_account(&caller, SubspaceMod::<T>::u64_to_balance(1_000_000_000_000_000).unwrap());
@@ -221,7 +221,7 @@ benchmarks! {
     remove_vote_proposal {
         let caller: T::AccountId = account("Alice", 0, 1);
         // Register Alice such that she has funds to vote
-        register_mock::<T>(caller.clone(), caller.clone(), 100000000000000u64, "test".as_bytes().to_vec())?;
+        register_mock::<T>(caller.clone(), caller.clone(), "test".as_bytes().to_vec())?;
 
         // Add Alice's funds to submit the proposal
         SubspaceMod::<T>::add_balance_to_account(&caller, SubspaceMod::<T>::u64_to_balance(1_000_000_000_000_000).unwrap());