+++ pallets/subspace/src/benchmarking.rs
@@ -10,25 +10,26 @@ use sp_std::vec::Vec;
 fn register_mock<T: Config>(
     key: T::AccountId,
     module_key: T::AccountId,
-    stake: u64,
     name: Vec<u8>,
 ) -> Result<(), &'static str> {
     let address = "test".as_bytes().to_vec();
     let network = "testnet".as_bytes().to_vec();
+
+    let enough_stake = 10000000000000u64;
     SubspaceMod::<T>::add_balance_to_account(
         &key,
-        SubspaceMod::<T>::u64_to_balance(stake + SubnetBurn::<T>::get() + 2000).unwrap(),
+        SubspaceMod::<T>::u64_to_balance(SubnetBurn::<T>::get() + enough_stake).unwrap(),
     );
     let metadata = Some("metadata".as_bytes().to_vec());
     SubspaceMod::<T>::register(
-        RawOrigin::Signed(key).into(),
+        RawOrigin::Signed(key.clone()).into(),
         network,
         name,
         address,
-        stake.into(),
-        module_key,
+        module_key.clone(),
         metadata,
     )?;
+    SubspaceMod::<T>::increase_stake(&key, &module_key, enough_stake);
     Ok(())
 }
 
@@ -44,10 +45,8 @@ benchmarks! {
         let module_key: T::AccountId = account("ModuleKey", 0, 2);
         let module_key2: T::AccountId = account("ModuleKey2", 0, 3);
 
-
-        let stake = 100000000000000u64;
-        register_mock::<T>(module_key.clone(), module_key.clone(), stake, "test".as_bytes().to_vec())?;
-        register_mock::<T>(module_key2.clone(), module_key2.clone(), stake, "test1".as_bytes().to_vec())?;
+        register_mock::<T>(module_key.clone(), module_key.clone(), "test".as_bytes().to_vec())?;
+        register_mock::<T>(module_key2.clone(), module_key2.clone(), "test1".as_bytes().to_vec())?;
         let netuid = SubspaceMod::<T>::get_netuid_for_name("testnet".as_bytes()).unwrap();
         let uids = vec![0];
         let weights = vec![10];
@@ -66,7 +65,7 @@ benchmarks! {
             &key,
             SubspaceMod::<T>::u64_to_balance(stake + 2000).unwrap(),
         );
-        register_mock::<T>(module_key.clone(), module_key.clone(), stake.clone(), "test".as_bytes().to_vec())?;
+        register_mock::<T>(module_key.clone(), module_key.clone(), "test".as_bytes().to_vec())?;
     }: add_stake(RawOrigin::Signed(key), module_key, stake)
 
     // 2
@@ -74,7 +73,7 @@ benchmarks! {
         let caller: T::AccountId = account("Alice", 0, 1);
         let module_key: T::AccountId = account("ModuleKey", 0, 2);
         let stake = 100000000000000u64;
-        register_mock::<T>(module_key.clone(), module_key.clone(), stake, "test".as_bytes().to_vec())?;
+        register_mock::<T>(module_key.clone(), module_key.clone(), "test".as_bytes().to_vec())?;
         let amount = 1000000000000;
         SubspaceMod::<T>::add_balance_to_account(
             &caller,
@@ -92,9 +91,8 @@ benchmarks! {
         let caller: T::AccountId = account("Alice", 0, 1);
         let module_key1: T::AccountId = account("ModuleKey1", 0, 2);
         let module_key2: T::AccountId = account("ModuleKey2", 0, 3);
-        let stake = 400000000000000u64;
-        register_mock::<T>(module_key1.clone(), module_key1.clone(), stake, "test".as_bytes().to_vec())?;
-        register_mock::<T>(module_key2.clone(), module_key2.clone(), stake, "test1".as_bytes().to_vec())?;
+        register_mock::<T>(module_key1.clone(), module_key1.clone(),"test".as_bytes().to_vec())?;
+        register_mock::<T>(module_key2.clone(), module_key2.clone(), "test1".as_bytes().to_vec())?;
         let module_keys = vec![module_key1, module_key2];
         let mut amounts = vec![100000000000000, 100000000000000];
         SubspaceMod::<T>::add_balance_to_account(
@@ -110,9 +108,8 @@ benchmarks! {
         let caller: T::AccountId = account("Alice", 0, 1);
         let module_key1: T::AccountId = account("ModuleKey1", 0, 2);
         let module_key2: T::AccountId = account("ModuleKey2", 0, 3);
-        let stake = 400000000000000u64;
-        register_mock::<T>(module_key1.clone(), module_key1.clone(), stake, "test".as_bytes().to_vec())?;
-        register_mock::<T>(module_key2.clone(), module_key2.clone(), stake, "test1".as_bytes().to_vec())?;
+        register_mock::<T>(module_key1.clone(), module_key1.clone(), "test".as_bytes().to_vec())?;
+        register_mock::<T>(module_key2.clone(), module_key2.clone(), "test1".as_bytes().to_vec())?;
         let module_keys = vec![module_key1.clone(), module_key2.clone()];
         let mut amounts = vec![100000000000000, 100000000000000];
         SubspaceMod::<T>::add_balance_to_account(
@@ -133,9 +130,8 @@ benchmarks! {
         let caller: T::AccountId = account("Alice", 0, 1);
         let module_key: T::AccountId = account("ModuleKey", 0, 2);
         let new_module_key: T::AccountId = account("NewModuleKey", 0, 3);
-        let stake = 100000000000000u64;
-        register_mock::<T>(module_key.clone(), module_key.clone(), stake, "test".as_bytes().to_vec())?;
-        register_mock::<T>(new_module_key.clone(), new_module_key.clone(), stake, "test1".as_bytes().to_vec())?;
+        register_mock::<T>(module_key.clone(), module_key.clone(), "test".as_bytes().to_vec())?;
+        register_mock::<T>(new_module_key.clone(), new_module_key.clone(), "test1".as_bytes().to_vec())?;
         let amount = 50000000000000;
         SubspaceMod::<T>::add_balance_to_account(
             &caller,
@@ -173,13 +169,12 @@ benchmarks! {
             &key,
             SubspaceMod::<T>::u64_to_balance(stake + SubnetBurn::<T>::get() + 2000).unwrap(),
         );
-    }: register(RawOrigin::Signed(key.clone()), "test".as_bytes().to_vec(), "test".as_bytes().to_vec(), "test".as_bytes().to_vec(), stake.into(), module_key.clone(), Some("metadata".as_bytes().to_vec()))
+    }: register(RawOrigin::Signed(key.clone()), "test".as_bytes().to_vec(), "test".as_bytes().to_vec(), "test".as_bytes().to_vec(),  module_key.clone(), Some("metadata".as_bytes().to_vec()))
 
     // 8
     deregister {
         let caller: T::AccountId = account("Alice", 0, 1);
-        let stake = 100000000000000u64;
-        register_mock::<T>(caller.clone(), caller.clone(), stake, "test".as_bytes().to_vec())?;
+        register_mock::<T>(caller.clone(), caller.clone(), "test".as_bytes().to_vec())?;
         let netuid = SubspaceMod::<T>::get_netuid_for_name("testnet".as_bytes()).unwrap();
     }: deregister(RawOrigin::Signed(caller), netuid)
 
@@ -191,7 +186,7 @@ benchmarks! {
     update_module {
         let caller: T::AccountId = account("Alice", 0, 1);
         let stake = 100000000000000u64;
-        register_mock::<T>(caller.clone(), caller.clone(), stake, "test".as_bytes().to_vec())?;
+        register_mock::<T>(caller.clone(), caller.clone(), "test".as_bytes().to_vec())?;
         let name = "updated_name".as_bytes().to_vec();
         let address = "updated_address".as_bytes().to_vec();
         let delegation_fee = Some(Percent::from_percent(5));
@@ -203,8 +198,7 @@ benchmarks! {
    // 10
    update_subnet {
         let caller: T::AccountId = account("Alice", 0, 1);
-        let stake = 100000000000000u64;
-        register_mock::<T>(caller.clone(), caller.clone(), stake, "test".as_bytes().to_vec())?;
+        register_mock::<T>(caller.clone(), caller.clone(), "test".as_bytes().to_vec())?;
         let netuid = SubspaceMod::<T>::get_netuid_for_name("testnet".as_bytes()).unwrap();
         let params = SubspaceMod::<T>::subnet_params(netuid);
     }: update_subnet(
@@ -230,4 +224,15 @@ benchmarks! {
         params.adjustment_alpha,
         params.min_immunity_stake
     )
+    // 11
+    delegate_rootnet_control {
+        use pallet_subnet_emission_api::SubnetConsensus;
+        let module_key: T::AccountId = account("ModuleKey", 0, 2);
+        let module_key2: T::AccountId = account("ModuleKey2", 0, 3);
+
+        register_mock::<T>(module_key.clone(), module_key.clone(), "test".as_bytes().to_vec())?;
+        register_mock::<T>(module_key2.clone(), module_key2.clone(), "test1".as_bytes().to_vec())?;
+        let netuid = SubspaceMod::<T>::get_netuid_for_name(b"testnet").unwrap();
+        T::set_subnet_consensus_type(netuid, Some(SubnetConsensus::Root));
+    }: delegate_rootnet_control(RawOrigin::Signed(module_key), module_key2)
 }